Time fix attempt

+++ src/WorldIMPLTalk.cpp
@@ -324,7 +324,7 @@ void World::sendIGTime(Player *cp) const {
     ServerCommandPointer cmd = std::make_shared<UpdateTimeTC>(
             static_cast<unsigned char>(getTime("hour")), static_cast<unsigned char>(getTime("minute")),
             static_cast<unsigned char>(getTime("day")), static_cast<unsigned char>(getTime("month")),
-            static_cast<short int>(getTime("year")));
+            static_cast<long>(getTime("year")));
     cp->Connection->addCommand(cmd);
 }
